Rework theme
Reworked the theme to be more flexible.  Additionally, discovered
how to make the site mobile-friendly.

This time, the navigation links are not fixed in place, as that
wasted reading space.

Decided to separate the home page from the about page.  Also
decided not to include as much common information per page - there
is no need for my name and face to be on every page.  Instead,
most pages will be more minimal.